+++ src/main/java/com/personalize/personalizeqa/server/impl/EntityServiceImpl.java
@@ -83,7 +83,7 @@ public class EntityServiceImpl extends ServiceImpl<EntityMapper, Entity> impleme
         queryWrapper.eq(Entity::getEntityName, entityName);
 
         // 使用 count 方法来获取匹配记录的数量
-        int count = count(queryWrapper);
+        long count = count(queryWrapper);
 
         // 如果记录数量大于 0，则说明存在
         return (count == 0);
